imported webdriver from selenium and beautifulsoup into the script to prepare for webscraping the nba official website.
set a path to webdriver and created instances for my webdriver so selenium can function properly once the script is ran.

hardcoded url of Boston Celtics nba.com page so that the webdriver can access it when called. added 'implicitly_wait' to allow javascript items to have the chance to load in before pulling.

created an html_content variable to store the html in once the javascript has finished running and then close the browser for selenium.

create a variable titled 'soup' for beautifulsoup and am ready to have it parse through the html.