added line that activates the soup find method to retrive all table rows from the html.
nested a for loop inside of an if statement to run a program where if there are any rows, then data cells will be pulled from that row. Tables will then be pulled based on how many data cells are in the row.

variables within the if statement is the data being pulled. I assigned a data cell at each index a specific variable and then stripped the html around the text of the cell so I could pull the necessary information.