rewrote code so that I could pull from the NBA.com site rather than the nba section of the espn site. Reasons for this include complications with duplicate players appearing if they had been traded during the season.
started on a team stats url. from there, found the class in the html that housed urls for specific teams. I ran a for loop that would allow the scraper to enter each team's url sequentially, and then would be able to pull the player names from each teams individual page.

From there, I also pulled each player's url for their personal page along with their name. While the name was sent to a mysql database for storage, the scraper used the urls to enter the player pages and get general, biographical information about each player to add to the database,