had to add webdrivers for javascript dropdown menus on the page. Issue was that random players automatically had their stats set for the preseason instead of the regular season.
imported 'webdriverwait' and 'expectedconditions' so that I would be able to create paths to the dropdown menus that changed the season types. I found that when I just clicked on the 'regular season' option once, the change would not always register. I added one for another table on the page and didn't change anything about it to ensure that when regular season was selected, the change would process.